volume-manifest-tool-31 Rename code to align with renamed project
