+++ service/README.md
@@ -1,10 +1,10 @@
 # TODO
-* copy sattva:~dev/volume-manifest-tool/etc up to Git
+* copy sattva:~dev/volume-manifest-builder/etc up to Git
 * document need credentials in /.aws (running as root)
 * document need manifest.sh installed from etc.
 * make sure all service/usr/local/bin files are executable by user `service`
 
-Define a service which launches on startup to run the volume manifest tool
+Define a service which launches on startup to run the volume manifest builder
 This service is built for a specific AWS AMI running Ubuntu (problems with Python 3.7 on ), but has also been tested on Debian 9 and 10 (stretch).
 ## Operating environment
 ### Prerequisites
@@ -13,14 +13,14 @@ This service is built for a specific AWS AMI running Ubuntu (problems with Pytho
 * Debian 9 or 10
 #### Software
 * python 3.7
-* volume_manifest_tool installed from pip for the global system. This should include its dependents (boto and others)
+* volume_manifest_builder installed from pip for the global system. This should include its dependents (boto and others)
 #### Service environment
 This service installation is built on the `systemctl` platform. The example `default.target` may be different on different Linuces.  On Debian 9, it is `multi-user.target`
 You will have to change the `manifest.service` file's  `[Install]` section's `WantedBy` property to match the host platform's target.
 #### Application environment
 * User: service
-* Folder ~service/volume-manifest-tool
-* Folder: /var/log/VolumeManifestTool - writable by `service`
+* Folder ~service/volume-manifest-builder
+* Folder: /var/log/VolumeManifestBuilder - writable by `service`
 * Environment variables required - See [Setting environment variable for service](https://serverfault.com/questions/413397/how-to-set-environment-variable-in-systemd-service). Create and populate these variables:
     * AO_AWS_SNS_TOPIC_ARN
 
@@ -31,7 +31,7 @@ No compilation of the service files is required.
 1. Become super user.
 1. Identify the file system that /etc is mounted on. You will need this because the `systemctl enable` command requires the service command files to be on the same partition as the /etc/systemd folder.
 1. Create a folder and move the $S_H usr/lib files (manifest.service) there (you dont have to preserve the usr/lib hierarchy. Anywhere is fine, but it should at least not e in your login folder.
-1. copy the $S_H/usr/local/bin/v-m-t folder to /usr/local/bin, so that you have /usr/local/bin/v-m-t
+1. copy the $S_H/usr/local/bin/v-m-b folder to /usr/local/bin, so that you have /usr/local/bin/v-m-b
 1. Run `sudo systemctl enable <path to>manifest.service. This should create a link in `/etc/systemd/system/<Install target you picked above>.d/` folder.
 
 This configures the system to launch the  /usr/local/bin/manifest-shell.sh on boot.  You can control and monitor the service using strandard `systemctl` commands.
