+++ v_m_b/detectgraynb.py
